task/: move most code to lib/
We have some libraries (GitHub API, AMQP, etc.) that have been living in
the task/ directory mostly because that's all that we had at the time
that code was written.  Let's move them to lib/ with the rest of the
newer stuff.

Update all of our internal users of those APIs but keep compatibility
shims around for github.py which is used from lcov.py in the cockpit
repository to post "This newly added line lacks test coverage" comments
and for distributed_queue and test_mock_server, both of which are used
from cockpituous.  We can update those later.
