+++ github-upload-secrets
@@ -39,7 +39,7 @@ from collections.abc import Mapping
 
 from nacl import encoding, public
 
-import task
+from lib import github
 
 
 def encrypt(public_key: str, secret_value: bytes) -> str:
@@ -81,7 +81,7 @@ def generate_ssh_key() -> tuple[bytes, bytes]:
 
 
 def upload_secrets(
-    api: task.github.GitHub, repo: str, env: str, secrets: Mapping[str, bytes], opts: argparse.Namespace
+    api: github.GitHub, repo: str, env: str, secrets: Mapping[str, bytes], opts: argparse.Namespace
 ) -> None:
     env_path = f"/repos/{repo}/environments/{env}"
 
@@ -107,7 +107,7 @@ def upload_secrets(
 
 
 def upload_deploy_key(
-    api: task.github.GitHub, repo: str, title: str, content: bytes, opts: argparse.Namespace
+    api: github.GitHub, repo: str, title: str, content: bytes, opts: argparse.Namespace
 ) -> None:
     keys_path = f"/repos/{repo.lstrip('+')}/keys"
 
@@ -133,7 +133,7 @@ def upload_deploy_key(
         api.post(keys_path, {'title': title, 'key': content.decode('ascii')})
 
 
-def do_directory(api: task.github.GitHub, repo: str, env: str, directory: str, opts: argparse.Namespace) -> None:
+def do_directory(api: github.GitHub, repo: str, env: str, directory: str, opts: argparse.Namespace) -> None:
     secrets = {}
     for entry in os.scandir(directory):
         with open(entry.path, 'rb') as f:
@@ -141,7 +141,7 @@ def do_directory(api: task.github.GitHub, repo: str, env: str, directory: str, o
     upload_secrets(api, repo, env, secrets, opts)
 
 
-def do_deploy_key(api: task.github.GitHub, deploy_to: str, deploy_from: str, opts: argparse.Namespace) -> None:
+def do_deploy_key(api: github.GitHub, deploy_to: str, deploy_from: str, opts: argparse.Namespace) -> None:
     repo, env, secret = deploy_from.rsplit('/', 2)
     public, private = generate_ssh_key()
 
@@ -150,7 +150,7 @@ def do_deploy_key(api: task.github.GitHub, deploy_to: str, deploy_from: str, opt
 
 
 def main() -> None:
-    api = task.github.GitHub()
+    api = github.GitHub()
 
     parser = argparse.ArgumentParser(description='Upload encrypted action secrets to GitHub')
     parser.add_argument('--repository', '-r', '--receiver', metavar="OWNER/REPO",
