+++ run-queue
@@ -31,10 +31,10 @@ from collections.abc import Sequence
 
 import pika
 
+from lib import distributed_queue
 from lib.directories import get_images_data_dir
 from lib.network import redhat_network
 from lib.stores import LOG_STORE
-from task import distributed_queue
 
 logging.basicConfig(level=logging.INFO)
