+++ tasks-container-update
@@ -27,7 +27,7 @@ from pathlib import Path
 import aiohttp
 from yarl import URL
 
-import task
+from lib import task
 from lib.aio.jsonutil import get_str, get_strv, typechecked
