+++ test/test_checklist.py
@@ -15,7 +15,7 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with Cockpit; If not, see <http://www.gnu.org/licenses/>.
 
-from task import github
+from lib import github
 
 
 def test_parse() -> None:
